add `variant_count` macro remove: unused dependencies
