+++ .github/workflows/rust.yml
@@ -11,6 +11,7 @@ env:
   CARGO_TERM_COLOR: always
   SCCACHE_GHA_ENABLED: true
   RUSTC_WRAPPER: "sccache"
+  CARGO_INCREMENTAL: 0
 
 concurrency:
   group: ci-${{ github.ref }}
