+++ core/Cargo.toml
@@ -22,7 +22,6 @@ level = { package = "inferno-level", path = "../level" }
 proto = { package = "inferno-proto", path = "../protocol" }
 replicator = { package = "inferno-repl", path = "../replication" }
 raknet = { package = "inferno-raknet", path = "../raknet" }
-macros = { package = "inferno-macros", path = "../macros" }
 
 console-subscriber = { version = "0.2.0", optional = true, features = ["parking_lot"] }
 
@@ -35,10 +34,8 @@ dashmap = "5.4.0"
 parking_lot = "0.12.1"
 lazy_static = "1.4.0"
 flate2 = "1.0.26"
-async-recursion = "1.0.2"
 serde = { version = "1.0.164", default-features = false }
 serde_json = { version = "1.0.111", features = ["preserve_order"] }
-clap = { version = "4.4.18", features = ["cargo", "std"], default-features = false }
 anyhow = { version = "1.0.76", features = ["backtrace"] }
 legion = "0.4.0"
 nohash-hasher = "0.2.0"
