+++ level/Cargo.toml
@@ -15,8 +15,6 @@ nbt = { package = "inferno-nbt", path = "../nbt" }
 proto = { package = "inferno-proto", path = "../protocol" }
 
 serde = { version = "1.0.194", features = ["derive"] }
-tokio-util = "0.7.8"
-tracing = "0.1.38"
 anyhow = { version = "1.0.79", features = ["backtrace"] }
 bytemuck = "1.13.1"
