+++ protocol/src/bedrock/input/mod.rs
@@ -1,0 +1,3 @@
+use util::glob_export;
+
+glob_export!(auth_input);
