+++ replication/Cargo.toml
@@ -8,6 +8,5 @@ proto = { package = "inferno-proto", path = "../protocol" }
 util = { package = "inferno-util", path = "../util" }
 
 anyhow = "1.0.79"
-tokio = { version = "1.35.1", features = ["macros", "rt"] }
 tracing = "0.1.40"
 fred = "7.1.2"
