+++ replication/src/lib.rs
@@ -46,7 +46,8 @@ impl Replicator {
             None,
             None,
         );
-        client.connect().await??;
+        client.connect();
+        client.wait_for_connect().await?;
 
         tracing::debug!("Replication layer created");
