+++ util/Cargo.toml
@@ -13,9 +13,7 @@ base64 = "0.21.7"
 cipher = "0.4.4"
 dashmap = "5.4.0"
 jsonwebtoken = "9.2.0"
-lazy_static = "1.4.0"
 num-traits = "0.2.15"
-parking_lot = "0.12.1"
 paste = "1.0.12"
 serde = "1.0.194"
 serde_json = "1.0.111"
